requirements.txt
boto3
zulip
pyyaml
